backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.12
alembic==1.15.2
pydantic-settings==2.9.1
python-multipart==0.0.20
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
celery==5.4.0
redis==5.2.1
aiofiles==24.1.0
patool==1.15.0
python-dotenv==1.1.0
httpx==0.28.1
openai==1.76.0
pytest==8.3.5
